receiver/requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pygame==1.9.3
pyparsing==2.2.0
six==1.10.0
Werkzeug==0.11.15
